attempt to remove the interface for ChatClient
