+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/ChatClientTest.kt
@@ -20,7 +20,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.util.Date
 
-internal class ChatClientImplTest {
+internal class ChatClientTest {
 
     companion object {
         val eventA = ConnectedEvent(EventType.HEALTH_CHECK, Date(), User(), "")
@@ -53,7 +53,7 @@ internal class ChatClientImplTest {
         )
 
         socket = FakeChatSocket()
-        client = ChatClientImpl(
+        client = ChatClient(
             config = config,
             api = mock(),
             socket = socket,
