+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/controllers/ChannelControllerImplSubscribeTest.kt
@@ -3,7 +3,6 @@ package io.getstream.chat.android.client.controllers
 import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.events.ChannelCreatedEvent
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.events.ConnectedEvent
