+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomain.kt
@@ -2,7 +2,6 @@ package io.getstream.chat.android.livedata
 
 import android.content.Context
 import androidx.lifecycle.LiveData
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.models.Config
 import io.getstream.chat.android.client.models.Mute
