+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomainFactory.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.os.Handler
 import android.os.Looper
 import androidx.room.Room
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.User
 
 internal class ChatDomainFactory {
