+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/ChannelRepository.kt
@@ -1,13 +1,11 @@
 package io.getstream.chat.android.livedata.repository
 
 import androidx.collection.LruCache
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.dao.ChannelDao
 import io.getstream.chat.android.livedata.entity.ChannelEntity
-import io.getstream.chat.android.livedata.extensions.isPermanent
 
 internal class ChannelRepository(
     var channelDao: ChannelDao,
