+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/ReactionRepository.kt
@@ -1,12 +1,10 @@
 package io.getstream.chat.android.livedata.repository
 
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.dao.ReactionDao
 import io.getstream.chat.android.livedata.entity.ReactionEntity
-import io.getstream.chat.android.livedata.extensions.isPermanent
 
 /**
  * We don't do any caching on reactions since usage is infrequent
