+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/utils/DefaultRetryPolicy.kt
@@ -1,6 +1,5 @@
 package io.getstream.chat.android.livedata.utils
 
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.errors.ChatError
 
 internal class DefaultRetryPolicy : RetryPolicy {
