+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/ChatDomainImplTest.kt
@@ -4,7 +4,6 @@ import android.os.Handler
 import androidx.arch.core.executor.testing.InstantExecutorExtension
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.User
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
