+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/utils/LiveDataTestUtil.kt
@@ -3,7 +3,6 @@ package io.getstream.chat.android.livedata.utils
 import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
-import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.socket.InitConnectionListener
